requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
langchain
langchain-community
langchain-openai
langchain-core
faiss-cpu
pypdf
tiktoken
python-multipart
scikit-learn
numpy
pycryptodome